Internal Changes.
PiperOrigin-RevId: 272257081

+++ models/official/detection/utils/dataloader_utils.py
@@ -22,10 +22,10 @@ from utils import input_utils
 def process_source_id(source_id):
   """Processes source_id to the right format."""
   if source_id.dtype == tf.string:
-    source_id = tf.cast(tf.string_to_number(source_id), tf.int32)
+    source_id = tf.cast(tf.string_to_number(source_id), tf.int64)
   with tf.control_dependencies([source_id]):
     source_id = tf.cond(tf.equal(tf.size(source_id), 0),
-                        lambda: tf.cast(tf.constant(-1), tf.int32),
+                        lambda: tf.cast(tf.constant(-1), tf.int64),
                         lambda: tf.identity(source_id))
   return source_id
